Integreated write_eeg_data into pipeline and scripts
